requirements.txt
accelerate==1.5.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
arclet-alconna==1.8.36
attrs==25.3.0
bbcode==1.1.0
beautifulsoup4==4.13.3
bs4==0.0.2
cbor==1.0.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
confz==2.1.0
cprint==1.2.2
datasets==3.4.0
dateparser==1.2.1
dill==0.3.8
filelock==3.18.0
filetype==1.2.0
FlagEmbedding==1.3.4
frozenlist==1.5.0
fsspec==2024.12.0
greenlet==3.1.1
grpcio==1.70.0
grpcio-tools==1.70.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
idna==3.10
ijson==3.3.0
inscriptis==2.5.3
ir_datasets==0.5.10
Jinja2==3.1.6
joblib==1.4.2
loguru==0.7.3
lxml==5.3.1
lz4==4.4.3
MarkupSafe==3.0.2
mpmath==1.3.0
msvc_runtime==14.42.34433
multidict==6.1.0
multiprocess==0.70.16
nepattern==0.7.7
networkx==3.4.2
numpy==2.2.3
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==24.2
pandas==2.2.3
peft==0.14.0
pil-utils==0.2.2
pillow==11.1.0
propcache==0.3.0
protobuf==5.29.3
psutil==7.0.0
pyarrow==19.0.1
pybind11==2.13.6
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pynng==0.8.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
qrcode==8.0
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==3.4.1
sentencepiece==0.2.0
six==1.17.0
skia-python==134.0b12
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
sympy==1.13.1
tarina==0.6.8
threadpoolctl==3.6.0
tokenizers==0.21.1
toml==0.10.2
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0
trec-car-tools==2.6
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
unlzw3==0.2.3
urllib3==2.3.0
warc3-wet==0.2.5
warc3-wet-clueweb09==0.2.5
watchdog==6.0.0
wcferry==39.4.2.2
win32_setctime==1.2.0
xxhash==3.5.0
yarl==1.18.3
zlib-state==0.1.9
